IssueID 999: fixed
git-svn-id: https://build.archivesportaleurope.net/svn/apex/branches/backend-project-1.5@2243 9fd842fd-667d-44da-918d-ff91420b3ae1

+++ Dashboard/src/main/java/eu/apenet/dashboard/services/ead/EadService.java
@@ -9,15 +9,13 @@ import java.util.Date;
 import java.util.List;
 import java.util.Properties;
 
-import eu.apenet.dashboard.manual.ExistingFilesChecker;
-import eu.apenet.persistence.vo.*;
-import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.Logger;
 
 import eu.apenet.commons.exceptions.APEnetRuntimeException;
 import eu.apenet.commons.types.XmlType;
 import eu.apenet.commons.utils.APEnetUtilities;
+import eu.apenet.dashboard.manual.ExistingFilesChecker;
 import eu.apenet.dashboard.security.SecurityContext;
 import eu.apenet.dashboard.services.ead.xml.XmlEadParser;
 import eu.apenet.dashboard.utils.ContentUtils;
@@ -29,8 +27,23 @@ import eu.apenet.persistence.dao.EseStateDAO;
 import eu.apenet.persistence.dao.QueueItemDAO;
 import eu.apenet.persistence.dao.UpFileDAO;
 import eu.apenet.persistence.factory.DAOFactory;
+import eu.apenet.persistence.vo.Ead;
+import eu.apenet.persistence.vo.Ese;
+import eu.apenet.persistence.vo.EseState;
+import eu.apenet.persistence.vo.EuropeanaState;
+import eu.apenet.persistence.vo.FindingAid;
+import eu.apenet.persistence.vo.HoldingsGuide;
+import eu.apenet.persistence.vo.IngestionprofileDefaultDaoType;
+import eu.apenet.persistence.vo.IngestionprofileDefaultExistingFileAction;
+import eu.apenet.persistence.vo.IngestionprofileDefaultNoEadidAction;
+import eu.apenet.persistence.vo.IngestionprofileDefaultUploadAction;
+import eu.apenet.persistence.vo.QueueAction;
+import eu.apenet.persistence.vo.QueueItem;
+import eu.apenet.persistence.vo.QueuingState;
+import eu.apenet.persistence.vo.SourceGuide;
+import eu.apenet.persistence.vo.UpFile;
+import eu.apenet.persistence.vo.ValidatedState;
 import eu.archivesportaleurope.persistence.jpa.JpaUtil;
-import java.util.Map;
 
 public class EadService {
 
@@ -220,14 +233,14 @@ public class EadService {
 			QueueItem queueItem = ead.getQueueItem();
 			ead.setQueuing(QueuingState.NO);
 			eadDAO.updateSimple(ead);
-			deleteFromQueueInternal(queueItem);
+			deleteFromQueueInternal(queueItem, true);
 		}
 		JpaUtil.commitDatabaseTransaction();
 	}
 
-	private static void deleteFromQueueInternal(QueueItem queueItem) throws IOException {
+	private static void deleteFromQueueInternal(QueueItem queueItem, boolean deleteUpFile) throws IOException {
 		UpFile upFile = queueItem.getUpFile();
-		if (upFile != null) {
+		if (upFile != null && deleteUpFile) {
 			String filename = APEnetUtilities.getDashboardConfig().getTempAndUpDirPath() + upFile.getPath() + upFile.getFilename();
 			File file = new File(filename);
 			File aiDir = file.getParentFile();
@@ -237,7 +250,7 @@ public class EadService {
 			}
 		}
 		DAOFactory.instance().getQueueItemDAO().deleteSimple(queueItem);
-		if (upFile != null) {
+		if (upFile != null && deleteUpFile) {
 			DAOFactory.instance().getUpFileDAO().deleteSimple(upFile);
 		}
 	}
@@ -275,6 +288,10 @@ public class EadService {
 
 	public static void deleteFromQueue(QueueItem queueItem) throws Exception {
 		SecurityContext.get().checkAuthorizedToManageQueue();
+		deleteFromQueue(queueItem, true);
+	}
+	
+	private static void deleteFromQueue(QueueItem queueItem, boolean deleteUpFile) throws Exception {
 		EadDAO eadDAO = DAOFactory.instance().getEadDAO();
 		JpaUtil.beginDatabaseTransaction();
 		Ead ead = queueItem.getEad();
@@ -282,7 +299,7 @@ public class EadService {
 			ead.setQueuing(QueuingState.NO);
 			eadDAO.updateSimple(ead);
 		}
-		deleteFromQueueInternal(queueItem);
+		deleteFromQueueInternal(queueItem, deleteUpFile);
 		JpaUtil.commitDatabaseTransaction();
 	}
 
@@ -427,15 +444,15 @@ public class EadService {
             //About EADID
             UpFile upFile = queueItem.getUpFile();
             LOGGER.info("Process queue item: " + queueItem.getId() + " " + queueItem.getAction() + ", upFile id: " + queueItem.getUpFileId() + "(" + xmlType.getName() + ")");
-
-            String eadid = ExistingFilesChecker.extractAttributeFromEad(APEnetUtilities.getDashboardConfig().getTempAndUpDirPath() + upFile.getPath() + upFile.getFilename(), "eadheader/eadid", null, true).trim();
-            if(StringUtils.isEmpty(eadid) || "empty".equals(eadid)) {
-                LOGGER.info("File has empty eadid!");
-                deleteFromQueue(queueItem);
+            String upFilePath = upFile.getPath() + upFile.getFilename();
+            String eadid = ExistingFilesChecker.extractAttributeFromEad(APEnetUtilities.getDashboardConfig().getTempAndUpDirPath() + upFilePath, "eadheader/eadid", null, true).trim();
+            if(StringUtils.isEmpty(eadid) || "empty".equals(eadid) ||  "error".equals(eadid)) {
                 if(ingestionprofileDefaultNoEadidAction.isRemove()) {
-                    deleteUpFile(upFile);
-                }
-                if(ingestionprofileDefaultNoEadidAction.isAddLater()) {
+                	LOGGER.info("File will be removed, because it does not have eadid: " + upFilePath);
+                	deleteFromQueue(queueItem, true);
+                }else {
+                	LOGGER.info("File will be processed manually later, because it does not have eadid: " + upFilePath);
+                	deleteFromQueue(queueItem, false);
                 }
             } else {
                 boolean continueTask = true;
@@ -474,7 +491,8 @@ public class EadService {
                             continueTask = false;
                         }
                     } else if(ingestionprofileDefaultExistingFileAction.isKeep()) {
-                        deleteUpFile(upFile);
+                    	LOGGER.info("File will be removed, because there is already one with the same eadid: " + upFilePath);
+                    	deleteFromQueue(queueItem, true);
                         continueTask = false;
                     }
                 } else {
@@ -533,15 +551,6 @@ public class EadService {
 		return queueAction;
 	}
 
-    private static void deleteUpFile(UpFile upFile) throws IOException {
-        JpaUtil.beginDatabaseTransaction();
-        DAOFactory.instance().getUpFileDAO().deleteSimple(upFile);
-        JpaUtil.commitDatabaseTransaction();
-        ContentUtils.deleteFile(APEnetUtilities.getDashboardConfig().getTempAndUpDirPath() + upFile.getPath() + upFile.getFilename());
-        File uploadDir = new File(APEnetUtilities.getDashboardConfig().getTempAndUpDirPath() + upFile.getPath());
-        if (uploadDir.listFiles().length == 0)
-            FileUtils.forceDelete(uploadDir);
-    }
 
     private static Ead doesFileExist(UpFile upFile, String eadid, XmlType xmlType) {
         return DAOFactory.instance().getEadDAO().getEadByEadid(xmlType.getClazz(), upFile.getAiId(), eadid);
@@ -675,7 +684,7 @@ public class EadService {
 			QueueItem queueItem = ead.getQueueItem();
 			ead.setQueuing(QueuingState.NO);
 			eadDAO.updateSimple(ead);
-			deleteFromQueueInternal(queueItem);
+			deleteFromQueueInternal(queueItem, true);
 			eads.remove(size - 1);
 		}
 		JpaUtil.commitDatabaseTransaction();
